[*] 功能优化
优化打包
优化图标
优化日志输出类名

+++ docs/build.sh
@@ -1,18 +1,17 @@
 #! /bin/bash
 
 # 配置参数
-
-pack=$3
+action=$3
 system=$2
 version=$1
 
-if [[ $version == "" ]] || [[ $system == "" ]]; then
-    echo "setting version and system：build.sh 1.0.0 [win|mac|linux] [pack]"
+# 判断条件
+if [[ $version == "" || $system == "" || $action == "" ]]; then
+    echo "setting version and system：build.sh 1.0.0 [win|mac|linux] [all|build|pack]"
     exit
 fi
 
 # 系统参数
-
 if [[ $system == "win" ]]; then
     # win:msi|exe
     icon="./docs/logo/logo.ico"
@@ -26,35 +25,32 @@ elif [[ $system == "linux" ]]; then
     icon="./docs/logo/logo.png"
     args="--type deb --linux-shortcut"
 else
-    echo "setting version and system：build.sh 1.0.0 [win|mac|linux] [pack]"
+    echo "setting version and system：build.sh 1.0.0 [win|mac|linux] [all|build|pack]"
     exit
 fi
 
 # 开始构建
-
 echo "Snail $version $system building..."
 
-# 删除文件
-
-mvn clean
-rm -rf ./build/
-
 # 编译项目
+if [[ $action == "all" || $action == "build" ]]; then
+    # 删除文件
+    mvn clean
+    rm -rf ./build/
+    # 编译项目
+    mvn package -P release -D javafx.platform=$system -D skipTests
+    mkdir -p ./build/snail/
+    # 拷贝文件
+    cp -vr ./snail-javafx/target/lib ./build/snail/
+    cp -vr ./snail-javafx/target/snail.javafx-${version}.jar ./build/snail/
+    tar -cvf ./build/snail-${system}-${version}-without-jre.tar -C ./build/ snail
+fi
 
-mvn package -P release -D javafx.platform=$system -D skipTests
-mkdir -p ./build/snail/
-cp -vr ./snail-javafx/target/lib ./build/snail/
-cp -vr ./snail-javafx/target/snail.javafx-${version}.jar ./build/snail/
-
-# 打包项目
-
-tar -cvf ./build/snail-${system}-${version}-without-jre.tar -C ./build/ snail
-
-if [[ $pack == "pack" ]]; then
-# 运行环境
-    jlink --add-modules "java.base,java.xml,java.desktop,java.scripting,jdk.unsupported" --output ./build/runtime
-# 打包软件
+# 打包项目：jlink|jpackage
+if [[ $action == "all" || $action == "pack" ]]; then
     jpackage \
+        --resource-dir ./docs/logo \
+        --file-associations ./docs/torrent.properties \
         --name snail \
         --app-version ${version} \
         --vendor acgist \
@@ -62,12 +58,13 @@ if [[ $pack == "pack" ]]; then
         --description "Acgist Snail Downloader" \
         --input ./build/snail/ \
         --main-jar snail.javafx-${version}.jar \
-        --runtime-image ./build/runtime \
+        --add-modules "java.base,java.xml,java.desktop,java.scripting,jdk.unsupported" \
         --icon ${icon} \
         --license-file ./LICENSE \
-        --file-associations ./docs/torrent.properties \
         --java-options "-server -Xms128m -Xmx256m -XX:NewRatio=2 -XX:SurvivorRatio=2 -Dfile.encoding=UTF-8" \
+        --about-url "https://gitee.com/acgist/snail" \
         --dest ./build/ \
+        --verbose \
         ${args}
 fi
